fix parameter checking for domain name/hosted id zone
the parameter checking is such that when no hosted id zone is present,
the stack will fail when the domain name is also not present. the same
is true in the reverse direction: when no domain name is specified, the
stack fails if no hosted id zone is present.

this change fixes the checks such that the assertions only come into
play when the initial field *is* present. for example, if the
hosted id zone *is* present, then the domain name must also be present.
similarly, when the domain name is present, the hosted id zone must
also be.